fix:add Constant file,remove console
