+++ src/components/Table/TablePropsTypes.tsx
@@ -1,8 +1,9 @@
+type TableValue= {
+  year: number;
+  valueEndOfYear: number;
+  interest: string;
+  annualInvestment: number;
+}
 export default interface TableProps {
-    values: {
-      year: number;
-      valueEndOfYear: number;
-      interest: string;
-      annualInvestment: number;
-    }[];
+    values:TableValue[];
 }
