+++ src/constants/typesOfInput.constant.ts
@@ -1,0 +1,13 @@
+enum InputValues {
+  INITIAL_INVESTMENT = "initialInvestment",
+  ANNUAL_INVESTMENT = "annualInvestment",
+  EXPECTED_RETURN = "expectedReturn",
+  DURATION = "duration",
+}
+enum InputNames {
+  INITIAL_INVESTMENT = "INITIAL INVESTMENT",
+  ANNUAL_INVESTMENT = "ANNUAL INVESTMENT",
+  EXPECTED_RETURN = "EXPECTED RETURN",
+  DURATION = "DURATION",
+}
+export {InputNames,InputValues};
